+++ Krowi_Databases/DbManager/DbManagerWPF/Model/PetFamily.cs
@@ -1,0 +1,18 @@
+﻿namespace DbManagerWPF.Model
+{
+    // <remarks>Needs to be in sync with the PetFamily table in the database.</remarks>
+    public enum PetFamily
+    {
+        Undefined = 0,
+        Aquatic = 1,
+        Beast = 2,
+        Critter = 3,
+        Dragonkin = 4,
+        Elemental = 5,
+        Flying = 6,
+        Humanoid = 7,
+        Magic = 8,
+        Mechanical = 9,
+        Undead = 10
+    }
+}