+++ Krowi_Databases/DbManager/DbManagerWPF/ViewModel/ExportViewModel.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Windows.Input;
 
 namespace DbManagerWPF.ViewModel
@@ -13,11 +14,16 @@ namespace DbManagerWPF.ViewModel
         #region Properties
         public ICommand ExportAchievementsCommand => new CommandHandler(() => ExportAchievements(), () => true);
 
+        public ICommand ExportNextPatchAchievementsCommand => new CommandHandler(() => ExportNextPatchAchievements(), () => true);
+
         public ICommand ExportCategoriesCommand => new CommandHandler(() => ExportCategories(), () => true);
 
         public ICommand ExportMapsCommand => new CommandHandler(() => ExportMaps(), () => true);
+
+        public ICommand ExportPetBattlesCommand => new CommandHandler(() => ExportPetBattles(), () => true);
         #endregion
 
+        #region ExportAchievements
         private void ExportAchievements()
         {
             var sb = new StringBuilder();
@@ -66,7 +72,64 @@ namespace DbManagerWPF.ViewModel
             using var file = new StreamWriter(@"../../../../../../Krowi_AchievementFilter/Data/ExportedAchievements.lua");
             file.WriteLine(sb.ToString());
         }
+        #endregion
 
+        #region ExportNextPatchAchievements
+        private void ExportNextPatchAchievements()
+        {
+            var sb = new StringBuilder();
+            sb.AppendLine($"-- [[ Exported at {DateTime.Now:yyyy-MM-dd HH-mm-ss} ]] --");
+            sb.AppendLine($"-- [[ This code is automatically generated as an export from ]] --");
+            sb.AppendLine($"-- [[ an SQLite database and is not meant for manual edit. ]] --");
+            sb.AppendLine("");
+            sb.AppendLine("-- [[ Namespaces ]] --");
+            sb.AppendLine("local _, addon = ...;");
+            sb.AppendLine("local data = addon.Data;");
+            sb.AppendLine("data.NextPatch = {};");
+            sb.AppendLine("local nextPatch = data.NextPatch;");
+            sb.AppendLine("");
+            sb.AppendLine("function nextPatch.Load(achievements)");
+            sb.AppendLineTabbed(1, "for i, _ in next, achievements do");
+            sb.AppendLineTabbed(2, "achievements[i] = nil;");
+            sb.AppendLineTabbed(1, "end");
+            sb.AppendLine("");
+
+            var categories = categoryDM.GetAll(true);
+            var category = categories.Single(x => x.ID == 799);
+
+            ExportNextPatchAchievements(sb, category);
+
+            sb.AppendLine("end");
+
+            using var file = new StreamWriter(@"../../../../../../Krowi_AchievementFilter/Data/ExportedNextPatch.lua");
+            file.WriteLine(sb.ToString());
+        }
+
+        private void ExportNextPatchAchievements(StringBuilder sb, Category category)
+        {
+            foreach (var subCategory in category.Children)
+            {
+                ExportNextPatchAchievements(sb, subCategory);
+
+                var achievements = achievementDM.GetWithCategory(subCategory);
+                foreach (var achievement in achievements)
+                {
+                    var achievementAGT = achievementDM.GetWithAGTID(achievement);
+                    var name = Regex.Replace(achievementAGT.Name, "^\"", "");
+                    name = Regex.Replace(name, "\"$", "");
+                    var description = Regex.Replace(achievementAGT.Description, "^\"", "");
+                    description = Regex.Replace(description, "\"$", "");
+                    description = Regex.Replace(description, "(\"){2,}", "\\\"");
+                    var rewardText = Regex.Replace(achievementAGT.RewardText, "^\"", "");
+                    rewardText = Regex.Replace(rewardText, "\"$", "");
+                    var line = $"achievements[{achievementAGT.ID}] = {{{achievementAGT.ID}, \"{name}\", {achievementAGT.Points}, \"{description}\", {achievementAGT.Flags}, {achievementAGT.IconFileID}, \"{rewardText}\"}};";
+                    sb.AppendLineTabbed(1, line);
+                }
+            }
+        }
+        #endregion
+
+        #region ExportCategories
         private void ExportCategories()
         {
             var sb = new StringBuilder();
@@ -86,9 +149,11 @@ namespace DbManagerWPF.ViewModel
             sb.AppendLineTabbed(1, "tinsert(table, value);");
             sb.AppendLineTabbed(1, "return value;");
             sb.AppendLine("end");
+            sb.AppendLine("");
             sb.AppendLine("local function AddC(c1, c2)");
             sb.AppendLineTabbed(1, "c1:AddCategory(c2)");
             sb.AppendLine("end");
+            sb.AppendLine("");
             sb.AppendLine("local function AddA(c, a)");
             sb.AppendLineTabbed(1, "c:AddAchievement(a)");
             sb.AppendLine("end");
@@ -143,7 +208,9 @@ namespace DbManagerWPF.ViewModel
             foreach (var child in category.Children)
                 ExportCategory(sb, child);
         }
+        #endregion
 
+        #region ExportMaps
         private void ExportMaps()
         {
             SortedDictionary<(UIMap UIMap, Function Function), List<Achievement>> keyValuePairs = new();
@@ -251,6 +318,117 @@ namespace DbManagerWPF.ViewModel
             foreach (var child in uiMap.Children)
                 LinkMaps(keyValuePairs, category, child, achievement);
         }
+        #endregion
+
+        #region ExportPetBattles
+        public void ExportPetBattles()
+        {
+            var sb = new StringBuilder();
+            sb.AppendLine($"-- [[ Exported at {DateTime.Now:yyyy-MM-dd HH-mm-ss} ]] --");
+            sb.AppendLine($"-- [[ This code is automatically generated as an export from ]] --");
+            sb.AppendLine($"-- [[ an SQLite database and is not meant for manual edit. ]] --");
+            sb.AppendLine("");
+            sb.AppendLine("-- [[ Namespaces ]] --");
+            sb.AppendLine("local _, addon = ...;");
+            sb.AppendLine("local objects = addon.Objects;");
+            sb.AppendLine("local data = addon.Data;");
+            sb.AppendLine("data.ExportedPetBattles = {};");
+            sb.AppendLine("local exportedPetBattles = data.ExportedPetBattles;");
+            sb.AppendLine("");
+            sb.AppendLine("local function AddCEL(m, t, e)");
+            sb.AppendLineTabbed(1, "m:AddChildExtLinkFull(t, e)");
+            sb.AppendLine("end");
+            sb.AppendLine("");
+            sb.AppendLine("local function AddCCEL(m, a, c, e)");
+            sb.AppendLineTabbed(1, "m:AddChildCritExtLinkFull(a, c, e)");
+            sb.AppendLineTabbed(1, "return m;");
+            sb.AppendLine("end");
+            sb.AppendLine("");
+            sb.AppendLine("function exportedPetBattles.Load(m)");
+            sb.AppendLineTabbed(1, "for i, _ in next, m do");
+            sb.AppendLineTabbed(2, "m[i] = nil;");
+            sb.AppendLineTabbed(1, "end");
+            sb.AppendLine("");
+            sb.AppendLineTabbed(1, "local url = \"https://www.wow-petguide.com/\";");
+            sb.AppendLine("");
+
+            var categories = categoryDM.GetAll(true);
+            foreach (var category in categories)
+                ExportPetBattles(sb, category);
+
+            sb.AppendLine("end");
+
+            using var file = new StreamWriter(@"../../../../../../Krowi_AchievementFilter/Data/ExportedPetBattles.lua");
+            file.WriteLine(sb.ToString());
+        }
+
+        private void ExportPetBattles(StringBuilder sb, Category category)
+        {
+            var achievements = achievementDM.GetWithCategory(category);
+            foreach (var achievement in achievements)
+            {
+                var criteria = petBattleLinksDM.GetWithAchievementID(achievement.ID);
+                if (criteria.Any())
+                {
+                    foreach (var criterion in criteria)
+                    {
+                        if (!string.IsNullOrEmpty(criterion.ExternalLink))
+                        {
+                            string externalLink;
+                            if (criterion.ExternalLink.Contains("https://www.wow-petguide.com"))
+                                externalLink = criterion.ExternalLink.Replace("https://www.wow-petguide.com/", "url .. \"");
+                            else
+                                externalLink = $"\"{criterion.ExternalLink}";
+
+                            var list = $"m[{achievement.ID}]";
+                            sb.AppendTabbed(1, $"{list} = objects.MenuItem:NewExtLink(addon.L[\"Xu-Fu's Pet Guides\"]");
+                            sb.Append(", ");
+                            sb.Append(externalLink);
+                            sb.AppendLine($"\"); -- {criterion.Name}");
+
+                            var subCriteria = petBattleLinksDM.GetWithParentID(criterion.ID);
+                            GetSubCriteria(sb, criterion.ID.Substring(1), subCriteria, list: list);
+                        }
+                    }
+                }
+            }
+
+            foreach (var child in category.Children)
+                ExportPetBattles(sb, child);
+        }
+
+        private void GetSubCriteria(StringBuilder sb, string achievementID, List<PetBattleLink> criteria, string assign = "", string list = "")
+        {
+            foreach (var criterion in criteria)
+            {
+                if (!string.IsNullOrEmpty(criterion.ExternalLink))
+                {
+                    var subCriteria = petBattleLinksDM.GetWithParentID(criterion.ID);
+
+                    var indent = 1;
+                    if (subCriteria.Any())
+                    {
+                        assign = $"temp{criterion.ID}";
+                        sb.AppendTabbed(1, $"local {assign} = ");
+                        indent = 0;
+                    }
+
+                    string externalLink;
+                    if (criterion.ExternalLink.Contains("https://www.wow-petguide.com"))
+                        externalLink = criterion.ExternalLink.Replace("https://www.wow-petguide.com/", "url .. \"");
+                    else
+                        externalLink = $"\"{criterion.ExternalLink}";
+
+                    if (criterion.ID.StartsWith("C"))
+                        sb.AppendLineTabbed(indent, $"AddCEL({list}, {criterion.Name}, {externalLink}\");");
+                    else
+                        sb.AppendLineTabbed(indent, $"AddCCEL({list}, {achievementID}, {criterion.CriteriaNumber}, {externalLink}\");");
+
+                    GetSubCriteria(sb, criterion.ID.Substring(1), subCriteria, list: assign);
+                }
+            }
+        }
+        #endregion
 
         private string TrimNils(string input)
         {